+++ CHANGELOG.md
@@ -29,3 +29,7 @@ All notable autonomous changes should be recorded here.
 ## 2026-07-12T09:00:00.000000+00:00
 
 - Added a simple Python utility to generate a 60x9 LED bitmap and added a test.
+
+## 2026-07-13T10:00:00.000000+00:00
+
+- Added a simple Python utility to generate a 60x9 LED bitmap and added a test.